requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pre-commit==4.0.1
python-dotenv==1.0.1
django-cors-headers==4.7.0
psycopg2-binary==2.9.9
